+++ ktx-location-services/src/androidTest/java/org/neidhardt/ktxlocation/services/GoogleLocationServiceTest.kt
@@ -5,15 +5,12 @@
 package org.neidhardt.ktxlocation.services
 
 import android.content.Context
-import android.location.LocationManager
 import androidx.test.filters.LargeTest
 import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
 import androidx.test.rule.GrantPermissionRule
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
-import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull