+++ CHANGELOG.md
@@ -7,10 +7,14 @@ SPDX-License-Identifier: CC-BY-4.0
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+## v0.2.0 - 2025-01-08
 ### Changed
 * Update build scripts to kotlin
 * Update build dependencies
 * Cleanup
+* Made play-services-location dependency compileTime only
+* Update play services to play-services-location:21.3.0
 
 ## v0.1.0 - 2022-07-07
 ### Added